+++ slb4j-config/slb4j-config-all/src/main/java/org/slb4j/config/all/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * All configuration formats for SLB4J.
+ */
+package org.slb4j.config.all;